Extract and test Assembly explorer
