+++ backend/src/Designer/Configuration/Extensions/ServiceCollectionExtensions.cs
@@ -1,22 +1,18 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
+using Altinn.Studio.Designer.Helpers;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
-using Microsoft.Extensions.DependencyModel;
 using Microsoft.Extensions.Options;
-using Newtonsoft.Json;
 
 namespace Altinn.Studio.Designer.Configuration.Extensions
 {
     public static class ServiceCollectionExtensions
     {
-        private const string AltinnAssemblyIdentifier = "Altinn.";
         public static IServiceCollection RegisterSettingsByBaseType<TMarker>(this IServiceCollection services, IConfiguration configuration)
         {
-            var typesToRegister = GetTypesAssignedFrom<TMarker>()
+            var typesToRegister = AltinnAssembliesScanner.GetTypesAssignedFrom<TMarker>()
                 .Where(type => !type.IsInterface && !type.IsAbstract);
 
             foreach (var configType in typesToRegister)
@@ -51,69 +47,6 @@ namespace Altinn.Studio.Designer.Configuration.Extensions
             services.Configure<TOption>(configuration.GetSection(sectionName));
             services.TryAddScoped(typeof(TOption), svc => ((IOptionsSnapshot<object>)svc.GetService(typeof(IOptionsSnapshot<TOption>))).Value);
         }
-
-        private static IEnumerable<Type> GetTypesAssignedFrom<TAssignedFrom>()
-        {
-            if (DependencyContext.Default == null)
-            {
-                throw new InvalidOperationException("PreserveCompilationContext should be set to true.");
-            }
-
-            return DependencyContext.Default
-                .RuntimeLibraries
-                .Where(IsAltinnLibrary)
-                .Where(IsLoadable)
-                .Select(library => Assembly.Load(new AssemblyName(library.Name)))
-                .GetTypesAssignedFrom<TAssignedFrom>();
-        }
-
-        private static bool IsLoadable(RuntimeLibrary library)
-        {
-            // Contains assembly
-            if (!library.RuntimeAssemblyGroups.Any())
-            {
-                return false;
-            }
-
-            try
-            {
-                Assembly.Load(new AssemblyName(library.Name));
-                return true;
-            }
-            catch (ArgumentNullException e) // On some machines there are some additional Runtime libraries with ".Reference" suffix. Skip them.
-            {
-                return false;
-            }
-        }
-
-        private static IEnumerable<Type> GetTypesAssignedFrom<TAssignedFrom>(this IEnumerable<Assembly> assemblies)
-        {
-            return assemblies.SelectMany(a => a.GetTypesAssignedFrom<TAssignedFrom>());
-        }
-
-        private static IEnumerable<Type> GetTypesAssignedFrom<TAssignedFrom>(this Assembly assembly)
-        {
-            return assembly.GetLoadableTypes().Where(type => typeof(TAssignedFrom).IsAssignableFrom(type) && type != typeof(TAssignedFrom));
-        }
-
-        private static IEnumerable<Type> GetLoadableTypes(this Assembly assembly)
-        {
-            try
-            {
-                return assembly.GetTypes();
-            }
-            catch (ReflectionTypeLoadException e)
-            {
-                return e.Types.Where(t => t != null);
-            }
-        }
-
-        private static bool IsAltinnLibrary(RuntimeLibrary library)
-        {
-            return library.Name.StartsWith(AltinnAssemblyIdentifier, StringComparison.OrdinalIgnoreCase)
-                || library.Dependencies.Any(d => d.Name.StartsWith(AltinnAssemblyIdentifier, StringComparison.OrdinalIgnoreCase));
-        }
-
 
     }
 }
